backend/requirements.txt
fastapi
uvicorn
python-multipart
yt-dlp
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
datasets
transformers
torch
torchvision
tqdm
pandas
opencv-python
Pillow
numpy
diskcache
aiohttp
